Populate DB with data for person model
