+++ models/models.go
@@ -1,0 +1,13 @@
+package models
+
+type Person struct {
+	ID          string `json:"id,omitempty"`
+	Firstname   string `json:"firstname,omitempty"`
+	Lastname    string `json:"lastname,omitempty"`
+	Contactinfo `json:"contactinfo,omitempty"`
+}
+type Contactinfo struct {
+	City    string `json:"city,omitempty"`
+	Zipcode string `json:"Zipcode,omitempty"`
+	Phone   string `json:"phone,omitempty"`
+}
