+++ person_data.json
@@ -1,0 +1,62 @@
+[
+  {
+    "_id": "5bbf590a4bfd65789b9512ce",
+    "firstname": "Church",
+    "lastname": "Mckinney",
+    "contactinfo": {
+      "city": "Ballico",
+      "zipcode": "938",
+      "phone": "+1 (854) 537-3050"
+    }
+  },
+  {
+    "_id": "5bbf590a42baed9048a7dc5c",
+    "firstname": "Katrina",
+    "lastname": "Dotson",
+    "contactinfo": {
+      "city": "Allendale",
+      "zipcode": "8203",
+      "phone": "+1 (908) 556-3374"
+    }
+  },
+  {
+    "_id": "5bbf590a17b1905e30735778",
+    "firstname": "Romero",
+    "lastname": "Kirk",
+    "contactinfo": {
+      "city": "Condon",
+      "zipcode": "8999",
+      "phone": "+1 (989) 538-2933"
+    }
+  },
+  {
+    "_id": "5bbf590a270a039e03ebc49d",
+    "firstname": "Jana",
+    "lastname": "Hess",
+    "contactinfo": {
+      "city": "Macdona",
+      "zipcode": "5917",
+      "phone": "+1 (876) 544-3171"
+    }
+  },
+  {
+    "_id": "5bbf590a43b70a45098d5a38",
+    "firstname": "Jennings",
+    "lastname": "Little",
+    "contactinfo": {
+      "city": "Kersey",
+      "zipcode": "4110",
+      "phone": "+1 (905) 599-3843"
+    }
+  },
+  {
+    "_id": "5bbf590afa26874dc83886a1",
+    "firstname": "Buckner",
+    "lastname": "Nelson",
+    "contactinfo": {
+      "city": "Connerton",
+      "zipcode": "1893",
+      "phone": "+1 (820) 452-3491"
+    }
+  }
+]
